Crate cc_datastore

This crate provides the feature of persistent storage for the transaction data expected to exist indefinitely.

Currently, it is implemented on the top of offchain storage, which is a persistent local storage of each node.

Structs

PermanentStorage

Permanent storage backed by offchain storage.

Enums

Error

Error type for datastore.

Traits

ChunkRootBackend

Backend for storing a map of (block_number, extrinsic_index) to chunk_root.

TransactionDataBackend

Permanent transaction data backend.